src/importer/requirements.txt
SQLAlchemy
pymssql
psycopg2

python_swiftclient==3.2.0
python-keystoneclient==3.8.0
